Feat: Actions permissions + general adjustments

+++ app/controllers/api/v1/reviews_controller.rb
@@ -6,8 +6,10 @@ module Api
       before_action :set_review, only: %i[show update destroy]
 
       def show
+        authorize @review
+
         if @review
-          render json: @review, status: :ok
+          render json: sanitize_review(@review), status: :ok
         else
           render json: { message: 'Data not available.' }, status: :unprocessable_entity
         end
@@ -15,23 +17,34 @@ module Api
 
       def create
         @review = Review.new(review_params)
+        @review.user_id = @current_user.id
+        @review.movie_id = Movie.find_by(uuid: review_params[:movie_id]).id
+
+        # TODO: Check logic to add to user's watched list
+        # @current_user.lists.where(type: 'watched') << Movie.find_by(uuid: review_params[:movie_id]).id
+
+        authorize @review
 
         if @review.save
-          render json: @review, status: :ok
+          render json: sanitize_review(@review), status: :ok
         else
           render json: { message: 'An error occurred.' }, status: :unprocessable_entity
         end
       end
 
       def update
+        authorize @review
+
         if @review.update(review_params)
-          render json: @review, status: :ok
+          render json: sanitize_review(@review), status: :ok
         else
           render json: { message: 'An error occurred.' }, status: :unprocessable_entity
         end
       end
 
       def destroy
+        authorize @review
+
         if @review.destroy
           render json: { message: 'Review deleted.' }, status: :ok
         else
@@ -42,11 +55,15 @@ module Api
       private
 
       def set_review
-        @review = Review.find(params[:id])
+        @review = policy_scope(Review).find_by!(uuid: params[:id])
+      end
+
+      def sanitize_review(data)
+        data.to_json(only: %i[uuid score note note_has_spoilers favorite created_at])
       end
 
       def review_params
-        params.require('review').permit(:id, :uuid, :score, :note, :note_has_spoilers, :favorite, :user_id, :movie_id)
+        params.require(:review).permit(:score, :note, :note_has_spoilers, :favorite, :movie_id, :uuid)
       end
     end
   end
